Add color_type TypedDict and update mask_colors loading
