+++ xrfeitoria/camera/camera_parameter.py
@@ -9,6 +9,7 @@ from xrprimer.data_structure.camera import PinholeCameraParameter
 from xrprimer.transform.convention.camera import convert_camera_parameter
 
 from ..data_structure.constants import PathLike, Vector
+from ..utils.converter import rotation_matrix
 
 
 class CameraParameter(PinholeCameraParameter):
@@ -251,129 +252,3 @@ class CameraParameter(PinholeCameraParameter):
 
     def __repr__(self) -> str:
         return f'CameraParameter(T={self.extrinsic_t}, convention="{self.convention}", world2cam={self.world2cam})'
-
-
-def rotation_matrix(angles: Tuple[float, float, float], order='xyz', degrees: bool = True) -> npt.NDArray[np.float32]:
-    """
-    Args:
-        angles (Tuple[float, float, float]): Rotation angles in degrees or radians.
-        order (str, optional): Rotation order. Defaults to 'xyz'.
-        degrees (bool, optional): Whether the input angles are in degrees. Defaults to True.
-    Returns:
-        ndarray: Rotation matrix 3x3.
-
-    Examples:
-        >>> rotation_matrix((0, 0, 0), order='xyz')
-
-    References:
-        https://programming-surgeon.com/en/euler-angle-python-en/
-    """
-    if degrees:
-        angles = np.deg2rad(angles)
-
-    theta1, theta2, theta3 = angles
-    c1 = np.cos(theta1)
-    s1 = np.sin(theta1)
-    c2 = np.cos(theta2)
-    s2 = np.sin(theta2)
-    c3 = np.cos(theta3)
-    s3 = np.sin(theta3)
-
-    if order == 'xzx':
-        matrix = np.array(
-            [
-                [c2, -c3 * s2, s2 * s3],
-                [c1 * s2, c1 * c2 * c3 - s1 * s3, -c3 * s1 - c1 * c2 * s3],
-                [s1 * s2, c1 * s3 + c2 * c3 * s1, c1 * c3 - c2 * s1 * s3],
-            ]
-        )
-    elif order == 'xyx':
-        matrix = np.array(
-            [
-                [c2, s2 * s3, c3 * s2],
-                [s1 * s2, c1 * c3 - c2 * s1 * s3, -c1 * s3 - c2 * c3 * s1],
-                [-c1 * s2, c3 * s1 + c1 * c2 * s3, c1 * c2 * c3 - s1 * s3],
-            ]
-        )
-    elif order == 'yxy':
-        matrix = np.array(
-            [
-                [c1 * c3 - c2 * s1 * s3, s1 * s2, c1 * s3 + c2 * c3 * s1],
-                [s2 * s3, c2, -c3 * s2],
-                [-c3 * s1 - c1 * c2 * s3, c1 * s2, c1 * c2 * c3 - s1 * s3],
-            ]
-        )
-    elif order == 'yzy':
-        matrix = np.array(
-            [
-                [c1 * c2 * c3 - s1 * s3, -c1 * s2, c3 * s1 + c1 * c2 * s3],
-                [c3 * s2, c2, s2 * s3],
-                [-c1 * s3 - c2 * c3 * s1, s1 * s2, c1 * c3 - c2 * s1 * s3],
-            ]
-        )
-    elif order == 'zyz':
-        matrix = np.array(
-            [
-                [c1 * c2 * c3 - s1 * s3, -c3 * s1 - c1 * c2 * s3, c1 * s2],
-                [c1 * s3 + c2 * c3 * s1, c1 * c3 - c2 * s1 * s3, s1 * s2],
-                [-c3 * s2, s2 * s3, c2],
-            ]
-        )
-    elif order == 'zxz':
-        matrix = np.array(
-            [
-                [c1 * c3 - c2 * s1 * s3, -c1 * s3 - c2 * c3 * s1, s1 * s2],
-                [c3 * s1 + c1 * c2 * s3, c1 * c2 * c3 - s1 * s3, -c1 * s2],
-                [s2 * s3, c3 * s2, c2],
-            ]
-        )
-    elif order == 'xyz':
-        matrix = np.array(
-            [
-                [c2 * c3, -c2 * s3, s2],
-                [c1 * s3 + c3 * s1 * s2, c1 * c3 - s1 * s2 * s3, -c2 * s1],
-                [s1 * s3 - c1 * c3 * s2, c3 * s1 + c1 * s2 * s3, c1 * c2],
-            ]
-        )
-    elif order == 'xzy':
-        matrix = np.array(
-            [
-                [c2 * c3, -s2, c2 * s3],
-                [s1 * s3 + c1 * c3 * s2, c1 * c2, c1 * s2 * s3 - c3 * s1],
-                [c3 * s1 * s2 - c1 * s3, c2 * s1, c1 * c3 + s1 * s2 * s3],
-            ]
-        )
-    elif order == 'yxz':
-        matrix = np.array(
-            [
-                [c1 * c3 + s1 * s2 * s3, c3 * s1 * s2 - c1 * s3, c2 * s1],
-                [c2 * s3, c2 * c3, -s2],
-                [c1 * s2 * s3 - c3 * s1, c1 * c3 * s2 + s1 * s3, c1 * c2],
-            ]
-        )
-    elif order == 'yzx':
-        matrix = np.array(
-            [
-                [c1 * c2, s1 * s3 - c1 * c3 * s2, c3 * s1 + c1 * s2 * s3],
-                [s2, c2 * c3, -c2 * s3],
-                [-c2 * s1, c1 * s3 + c3 * s1 * s2, c1 * c3 - s1 * s2 * s3],
-            ]
-        )
-    elif order == 'zyx':
-        matrix = np.array(
-            [
-                [c1 * c2, c1 * s2 * s3 - c3 * s1, s1 * s3 + c1 * c3 * s2],
-                [c2 * s1, c1 * c3 + s1 * s2 * s3, c3 * s1 * s2 - c1 * s3],
-                [-s2, c2 * s3, c2 * c3],
-            ]
-        )
-    elif order == 'zxy':
-        matrix = np.array(
-            [
-                [c1 * c3 - s1 * s2 * s3, -c2 * s1, c1 * s3 + c3 * s1 * s2],
-                [c3 * s1 + c1 * s2 * s3, c1 * c2, s1 * s3 - c1 * c3 * s2],
-                [-c2 * s3, s2, c2 * c3],
-            ]
-        )
-
-    return matrix
